+++ GamingManagement-service/src/main/java/cz/fi/muni/pa165/yellow_yak/facade/TeamFacadeImpl.java
@@ -13,6 +13,8 @@ import javax.inject.Inject;
 import java.util.List;
 
 /**
+ * Implementation for team facade
+ *
  * @author Matej Knazik
  */
 @Service
@@ -21,10 +23,10 @@ public class TeamFacadeImpl implements TeamFacade {
 
     final static Logger log = LoggerFactory.getLogger(TeamFacadeImpl.class);
 
-    @Inject
+    @Autowired
     private TeamService teamService;
 
-    @Inject
+    @Autowired
     private BeanMappingService beanMappingService;
 
     @Autowired
@@ -38,23 +40,32 @@ public class TeamFacadeImpl implements TeamFacade {
         if (name == null) {
             throw new IllegalArgumentException("name cannot be null");
         }
+        if (name.isEmpty()) {
+            throw new IllegalArgumentException("invalid argument");
+        }
         log.info("create team, name = {}", name);
         return beanMappingService.mapTo(teamService.create(name), TeamDTO.class);
     }
 
     @Override
-    public void remove(Long teamId) {
+    public boolean remove(Long teamId) {
         if (teamId == null) {
             throw new IllegalArgumentException("teamId cannot be null");
         }
+        if (teamId <= 0) {
+            throw new IllegalArgumentException("id cannot be zero or negative value");
+        }
         log.info("removing team, id = {}", teamId);
-        teamService.remove(teamId);
+        return teamService.remove(teamId);
     }
 
     @Override
     public TeamDTO findById(Long teamId) {
         if (teamId == null) {
             throw new IllegalArgumentException("teamId cannot be null");
+        }
+        if (teamId <= 0) {
+            throw new IllegalArgumentException("id cannot be zero or negative value");
         }
         log.info("finding team by ID, id = {}", teamId);
         return beanMappingService.mapTo(teamService.findById(teamId), TeamDTO.class);
@@ -64,6 +75,9 @@ public class TeamFacadeImpl implements TeamFacade {
     public List<TeamDTO> findByName(String name) {
         if (name == null) {
             throw new IllegalArgumentException("name cannot be null");
+        }
+        if (name.isEmpty()) {
+            throw new IllegalArgumentException("invalid argument");
         }
         log.info("finding teams by name, name = {}", name);
         return beanMappingService.mapTo(teamService.findByName(name), TeamDTO.class);
